Added PointingFilter
Not sure I really need it, because it's just an alternative to doing
touchpoint sprite feedback with the qml runtime, as I began to describe
in https://www.qt.io/blog/qml-runtime-features-part-1 and still didn't
publish the second part yet. But sometimes there are reasons to run a
presentation with a custom executable instead of with qml.